Grad 2/Backend/Face_recognition_in_real-time/requirements.txt
opencv-python
face_recognition
numpy
pymongo
python-dotenv